:sparkles: feat(yoz): add uri module with URI parsing and manipulation utilities
Add comprehensive URI handling module to yoz with the following functions:
- parse/build: parse URI into parts and rebuild from components
- basename/extname/pathname: extract filename, extension, and path
- parent: get parent directory URI
- protocol/hash: extract protocol scheme and hash fragment
- join/relative: path joining and relative path computation
- normalize: normalize URI paths
- encode/decode: percent-encoding for URI components
- is_data_uri/validate: URI type checking and validation

+++ rust/yoz/src/dict/search.rs
@@ -127,7 +127,7 @@ fn collect_fuzzy_candidates(
             if line_match.score == 0 {
                 return None;
             }
-            let index = line_match.lnum.checked_sub(1)? as usize;
+            let index = line_match.lnum.checked_sub(1)?;
             if index >= group.tails.len() {
                 return None;
             }
